Support "openssl verify" to check for a trusted CA
The Python code I'd originally written to validate that a received
certificate has a verification path to a trusted CA only works with very
new versions of the OpenSSL.crypto module.  So new that the required
support (X509Store.load_locations) isn't even present in the Debian 10
python3-openssl package.

To make the check that the certificate was issued by a trusted CA work
in the (common!) case where the Python support isn't available, I added
a second implementation that invokes "openssl verify" in a subprocess.

And now that we can successfully perform this verification, I've
uncommented the code triggering it.