Avoid death by ProcessLookupError
There was a race condition handling timed-out communication with a
reload subprocess.  When subprocess.communicate times out, we try to
kill the subprocess.  But the subprocess could go ahead and exit on its
own before we are able to kill it.  Killing a non-existant process
raises ProcessLookupError.  Resolve by handling ProcessLookupError and
doing nothing-- it means the subprocess is done, which is what we want.