Use tuple of potential SSL Context methods
Older versions of the OpenSSL.SSL module may not support some of the
method constants available in newer versions of the module.  Most
notably, I saw a version that did not have the TLS_METHOD constant that
the documentation recommends new code use.

Rather than hard code logic that tries TLS_METHOD and then falls back to
TLSv1_METHOD, iterate a tuple with allowable methods ordered by
preference.  Construct the SSL context used to verify a certificate and
key match with the first method in the tuple that is available.