Fix verify_certificate_matches_key failures
It turns out that when an OpenSSL.SSL.Context is given a key that
mismatches a previously provided certificate, it raises an exception
immediately, not when check_privatekey is called.  So I moved the
use_privatekey call into the try block.  I also added logic that
attempts to confirm that, if an OpenSSL.SSL.Error was raised, it was
caused by a key mismatch.  Only then is BadCertificateException raised.