Tests for install_files failure recovery
These tests inject errors into the write and rename operations performed
while installing files and confirm that the target and back-up files
remain in a reasonable state.

Currently three of the tests are skipped because cert_receive.py does
not properly handle removal of back-up files associated with target
files that were not actually installed because of the error injection.