Use consistent order for install_files operations
To make tests easier to understand, adjust install_files so that each
stage first works on an applicable certificate file, then an applicable
intermediates bundle file, then an applicable key file.  This makes it
so that failing the nth write_one_file call and failing the nth rename
call impact installation of the same thing.  E.g., with no bundling, the
certificate is handled by the 1st calls, the intermediate bundle for the
2nd calls, etc.