Unnecessary back-up files are not inconsistencies
When installing to a file that already exists with different content,
cert_receieve does the following:

 1. Creates a tmpfile with new content
 2. Backs up the existing file's content
 3. Renames the tmpfile on top of the existing file

If step 3 fails, as part of the recovery process, we try to remove the
back-up file created by step 2.  If we cannot unlink it during the
recovery process, we leave behind a consistent back-up file identical to
the target file.  In this case, both the target and back-up files are
semantically valid.  And, assuming no other recovery errors, the target
certificate, intermediates bundle, and key files will all preserve the
content they had before the cert_receieve invocation.

As a consequence, we should not indicate there's a potential
inconsistency if we are unable to remove a successfully created back-up
file.