Three fixes for test_verify_trust_chain_...
  1. Apparently intermediate certificates need to be marked as CAs in
     their basic constraints.
  2. The verifier routines expect certificates to be ordered so that
     the app certificate is first and the intermediate issued by a known
     CA is last.  The test cases were providing them in reversed order.
  3. The wrong configuration (lacking the necessary ca_file setting) was
     used in test_verify_trust_chain_negative_2.

The positive test cases didn't catch bug (1) because of bug (2); the
verifier happily indicated that the CA-signed intermediate was signed by
a known CA, because it was.

The ..._negative_2 test case is using a validation chain that looks
like: App -> Inter1 -> Inter2 -> CA but intentionally is not passing one
of the intermediate certificates to the verifier.  Before I found bug
(3), I thought the reason why it didn't catch bug (2) was because the
suite was only covering the case where Inter2 was missing-- and since
Inter1 is not signed by the CA, verification would still fail, even
though the App certificate was not verified as intended.  It turned out
I was wrong-- but I didn't figure that out until after I'd already
written a test case leaving out the the other intermediate certificate.
Since this is clearly tricky, I figured might as well keep it; it's the
new ..._negative_3 test case.