Add Python project setuptools packaging/build
The initial rationale for this change was that use of a Python build
system would more easily support installing cert_receive.py as a
package/module, which can then be pre-compiled, improving performance.
There's a pybuild debhelper tool that eases building Debian packages
from Python package source distributions and I thought it would be the
easiest route to getting a cert_receive package installed properly from
a Python perspective.

In retrospect I've concluded that the Python build/install process is
too limited to use exclusively.  There seems to be no standard or
recommended way to install non-Python files outside of the Python
package's directory.  And cert-propagation includes several such things:
man pages, certbot_send.sh, and perhaps eventually configuration
templates/examples.  Executable script generation of scripts with a
<foo>.py name is troublesome because they include their parent directory
in sys.path, and hence then cannot import a <foo> module installed
elsewhere.  Finally, the generated scripts also do not work with
packages installed in a private location.

Despite the issues, I'm still adding the Python packaging because it is
relatively simple and I think there are a couple benefits.  One is that
it handles properly generating Python package metadata.  A second is
that it makes cert_receive more accessible and easy to use for people
familiar with the Python ecosystem but who may not use Debian and lack
experience installing non-Python packages from source.